+++ packages/xrpl/src/ECDSA.ts
@@ -1,6 +1,7 @@
 enum ECDSA {
   ed25519 = 'ed25519',
   secp256k1 = 'ecdsa-secp256k1',
+  sm2 = 'sm2'
 }
 
 export default ECDSA