+++ packages/xrpl/src/models/common/index.ts
@@ -1,7 +1,7 @@
 export type LedgerIndex = number | ('validated' | 'closed' | 'current')
 
 interface XRP {
-  currency: 'XRP'
+  currency: 'HWA'
 }
 
 interface IssuedCurrency {