+++ packages/xrpl/src/sugar/balances.ts
@@ -115,7 +115,7 @@ async function getBalances(
         formatBalances(response.result.lines),
       )
       if (xrpBalance !== '') {
-        balances.push({ currency: 'XRP', value: xrpBalance })
+        balances.push({ currency: 'HWA', value: xrpBalance })
       }
       balances.push(...accountLinesBalance)
     },